+++ flutter_tide/lib/flutter_tide.dart
@@ -4,3 +4,4 @@ export 'package:tide/tide.dart';
 export 'package:provider/provider.dart';
 
 export 'src/provider.dart';
+export 'src/persisted_provider.dart';