+++ tide/lib/tide.dart
@@ -5,3 +5,5 @@ export 'package:meta/meta.dart';
 export 'src/action.dart';
 export 'src/store.dart';
 export 'src/service.dart';
+export 'src/persistence/persistence.dart';
+export 'src/persistence/json_file.dart';