+++ flutter_tide/CHANGELOG.md
@@ -1,3 +1,19 @@
+## 0.7.2
+
+- Renamed `initialState` to `initialStore` for persisted state provider.
+
+## 0.6.5
+
+- Fixed state restoration.
+
+## 0.6.1
+
+- Updated tide dependency.
+
+## 0.5.1
+
+- Added persisted store provider.
+
 ## 0.4.0
 
 - Added service dependencies.